refactor: inline PlantState into chlorine benchmark and format cleanup
Move PlantState definition from plant_state.py into chlorine.py (only
consumer). Delete plant_state.py. Minor formatting (import order, blank
lines) from linter.
