+++ libs/process-control/src/process_control/chemistry/ph_model.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass
 
-import jax.numpy as jnp
 import jax
+import jax.numpy as jnp
 
 
 @dataclass(frozen=True)
